polkavm-linker: Add a DecoderConfig struct
We plan to extend decoder to include Zbb instruction set, while keeping
the backward compatibility of the decoder in mind, we should add a flag
to turn on this instruction set.

Since we already have a flag for 64-bit, it will soon become ugly to keep
passing a bunch of parameters to the decoder. Let's create a DecoderConfig
to bundle the flags for the decoder.

This patch only aims to refactor code, and no functional change is intended.

Signed-off-by: Aman <aman@parity.io>